feat(pancake squad): Lift up state refactor
